mcp/clients/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "MCP Python Client"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "groq>=0.30.0",
    "mcp>=1.11.0",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
